+++ WalletWasabi.Tests/UnitTests/BitcoinCore/NodeBuildingTests.cs
@@ -22,5 +22,19 @@ namespace WalletWasabi.Tests.UnitTests.BitcoinCore
 				await coreNode.StopAsync();
 			}
 		}
+
+		[Fact]
+		public async Task DataDirDiffersAsync()
+		{
+			var coreNodes = await Task.WhenAll(CoreNode.CreateAsync(additionalFolder: "0"), CoreNode.CreateAsync(additionalFolder: "1"));
+			try
+			{
+				Assert.NotEqual(coreNodes[0].DataDir, coreNodes[1].DataDir);
+			}
+			finally
+			{
+				await Task.WhenAll(coreNodes[0].StopAsync(), coreNodes[1].StopAsync());
+			}
+		}
 	}
 }
