+++ WalletWasabi/Helpers/EnvironmentHelpers.cs
@@ -131,6 +131,27 @@ namespace WalletWasabi.Helpers
 			return directory;
 		}
 
+		// This method removes the path and file extension.
+		//
+		// Given Wasabi releases are currently built using Windows, the generated assymblies contain
+		// the hardcoded "C:\Users\User\Desktop\WalletWasabi\.......\FileName.cs" string because that
+		// is the real path of the file, it doesn't matter what OS was targeted.
+		// In Windows and Linux that string is a valid path and that means Path.GetFileNameWithoutExtension
+		// can extract the file name but in the case of OSX the same string is not a valid path so, it assumes
+		// the whole string is the file name.
+		public static string ExtractFileName(string callerFilePath)
+		{
+			var lastSeparatorIndex = callerFilePath.LastIndexOf("\\");
+			if (lastSeparatorIndex == -1)
+			{
+				lastSeparatorIndex = callerFilePath.LastIndexOf("/");
+			}
+
+			lastSeparatorIndex++;
+			var fileNameWithoutExtension = callerFilePath.Substring(lastSeparatorIndex, callerFilePath.Length - lastSeparatorIndex - ".cs".Length);
+			return fileNameWithoutExtension;
+		}
+
 		/// <summary>
 		/// Executes a command with bash.
 		/// https://stackoverflow.com/a/47918132/2061103
